My Commit 19/11

+++ QLDHCDAPI/QLDHCDAPI/Models/CT_BAUBKS.cs
@@ -17,7 +17,7 @@ namespace QLDHCDAPI.Models
         public int MACT { get; set; }
         public string MABKS { get; set; }
         public string MANGUOIBAU { get; set; }
-        public int SLPHIEUBAU { get; set; }
+        public long SLPHIEUBAU { get; set; }
         public string HINHTHUCBAU { get; set; }
         public Nullable<System.DateTime> THOIGIANBAU { get; set; }
         public bool LAHOPLE { get; set; }
